Added routing algorithm using unification.
Both named parameters and splats are allowed. From any URI
a path from Nodes is built. If route is added, the path is
merged into root Node. If matching, an URI to match is
converted to Nodes path, than unification happens.

Memory managment is poor right now and the routing process
is not yet integrated.